split components into seperate files
